Added musician components

+++ src/app/components/musician-detail/musician-detail.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs/internal/Observable';
+import { Group } from 'src/app/models/Group';
 import { Musician } from 'src/app/models/Musician';
+import { AuthService } from 'src/app/services/auth.service';
+import { GroupService } from 'src/app/services/group.service';
 import { MusicianService } from 'src/app/services/musician.service';
 
 @Component({
@@ -10,12 +13,20 @@ import { MusicianService } from 'src/app/services/musician.service';
   styleUrls: ['./musician-detail.component.css']
 })
 export class MusicianDetailComponent implements OnInit {
-  public musicianId:number
+  public musicianId:number;
   public musician: Musician;
-  public result: any;
+  public groupId: number;
+  public group: Group;
+
+  public get isAdmin(): boolean{
+    return this.authService.isAdmin();
+  }
 
   constructor(private musicianService:MusicianService,
-              private avRoute: ActivatedRoute){
+              private groupService:GroupService,
+              private avRoute: ActivatedRoute,
+              private route: Router,
+              private authService: AuthService){
     if (this.avRoute.snapshot.params['id']) {
       this.musicianId = this.avRoute.snapshot.params['id'];
     }
@@ -25,18 +36,29 @@ export class MusicianDetailComponent implements OnInit {
     this.loadMusician();
   }
 
+  loadGroup(){
+    this.groupService.getGroup(this.musician.groupId)
+    .subscribe(res => {
+      this.group = res;
+    })
+  }
+
   loadMusician(){
     this.musicianService.getMusician(this.musicianId)
     .subscribe(res => {
       this.musician = res;
+      this.loadGroup();
     })
   }
 
   deleteMusician(){
-    this.musicianService.deleteMusician(this.musicianId)
-    .subscribe(res => {
-      this.result = res;
-    })
+    const ans = confirm('Are you shure you want to delete this Musician?');
+    if (ans) {
+      this.musicianService.deleteMusician(this.musicianId)
+      .subscribe(res => {
+        this.route.navigate(["/"]);
+      })
+    }
   }
 
 }
